requirements.txt
ipython==8.4.0
isort==5.10.1
matplotlib==3.1.2
natsort==8.1.0
numpy==1.17.4
pandas==0.25.3
Pillow==9.1.1
requests==2.22.0
roc_utils==0.2.2
scikit_image==0.16.2
scikit_learn==1.1.1
skimage==0.0
streamlit==1.9.0
tensorflow==2.9.1
tensorflow_gpu==2.5.0
torch==1.9.0
tqdm==4.64.0
